Handover

+++ src/main/java/com/example/pets_backend/service/UserService.java
@@ -1,7 +1,5 @@
 package com.example.pets_backend.service;
 
-import com.example.pets_backend.entity.Event;
-import com.example.pets_backend.entity.Task;
 import com.example.pets_backend.entity.User;
 import com.example.pets_backend.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -33,40 +31,52 @@ public class UserService implements UserDetailsService {
     private final SendMailService sendMailService;
     private final SchedulerService schedulerService;
 
+    /**
+     * Load the user by email
+     * @param email email
+     * @return a Spring in-built User object created by the user's email and password
+     * @throws UsernameNotFoundException when email not found or not verified
+     */
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
         User user = userRepo.findByEmail(email);
-        checkUserForLogin(user);
+        if (user == null) {
+            throw new UsernameNotFoundException("User not found");
+        }
+        if (!user.isEmail_verified()) {
+            throw new UsernameNotFoundException("Email not verified");
+        }
         return new org.springframework.security.core.userdetails.User(email, user.getPassword(), new ArrayList<>());
     }
 
-    private void checkUserForLogin(User user) {
-        if (user == null) {
-            throw new EntityNotFoundException("User not found");
-        }
-        if (!user.isEmail_verified()) {
-            throw new EntityNotFoundException("Email not verified");
-        }
-    }
-
+    /**
+     * Check for duplicate email => encrypt password => store the user into repository
+     * @param user User
+     * @return user
+     */
     public User save(User user) {
         String email = user.getEmail();
-        if (userRepo.findByEmail(email) != null) {
-            log.error("Duplicate email '" + email + "'");
+        if (userRepo.existsByEmail(email)) {
             throw new DuplicateKeyException(("Duplicate email '" + email + "'"));
-        } else {
-            log.info("New user '{}' saved into database", user.getUid());
         }
+        // encrypt the password before stored into repository
         user.setPassword(passwordEncoder.encode(user.getPassword()));
         return userRepo.save(user);
     }
 
+    /**
+     * Build a plain text email with the verification link and send to the user's email immediately.
+     * Do NOT throw exceptions if the mail-sending job failed
+     * @param user user
+     */
     public void sendVerifyEmail(User user) {
         String email = user.getEmail();
-        String token = user.getVerify_token();
+        String token = user.getVerify_token();  // Get the automatically generated verify token
+        // verifyUrl example: https://pets-tracking.azurewebsites.net/#/user/verify?email=petpocket@gmail.com&token=BGkrYVcUe6VWKPKVYCi7k
         String verifyUrl = WEB_PREFIX + "#/user/verify?" +
                 "email=" + email + "&" +
                 "token=" + token;
+        // Plain-text content of the verification email
         String text = "Hi " + user.getFirstName() + ", \n\n" +
                 "Click the following link to verify your email: \n" +
                 verifyUrl + "\n\n";
@@ -79,10 +89,26 @@ public class UserService implements UserDetailsService {
         }, LocalDateTime.now());
     }
 
+    /**
+     * Find the user by uid, throw exception if user not found
+     * @param uid uid
+     * @return user
+     */
     public User findByUid(String uid) {
         User user = userRepo.findByUid(uid);
-        checkUserInDB(user, uid);
+        if (user == null) {
+            throw new EntityNotFoundException("User not found");
+        }
         return user;
+    }
+
+    /**
+     * Find the User by email (do NOT throw exceptions if not found)
+     * @param email email
+     * @return the user registered with the email; null if not found
+     */
+    public User findByEmail(String email) {
+        return userRepo.findByEmail(email);
     }
 
     public List<User> findAll() {
@@ -90,42 +116,27 @@ public class UserService implements UserDetailsService {
         return userRepo.findAll();
     }
 
+    /**
+     * Delete the user by uid, throw exception if user not found
+     * @param uid uid
+     */
     public void deleteByUid(String uid) {
-        User user = userRepo.findByUid(uid);
-        checkUserInDB(user, uid);
-        userRepo.deleteById(uid);
-        log.info("User '{}' deleted from database", uid);
-    }
-
-    public void deleteByEmail(String email) {
-        User user = userRepo.findByEmail(email);
-        checkUserInDB(user, email);
-        userRepo.deleteByEmail(email);
-        log.info("User with email '{}' deleted from database", email);
-    }
-
-    public User findByEmail(String email) {
-        return userRepo.findByEmail(email);
-    }
-
-    public Event getEventByUidAndEventId(String uid, String eventId) {
-        User user = userRepo.findByUid(uid);
-        checkUserInDB(user, uid);
-        return user.getEventByEventId(eventId);
-    }
-
-    public Task getTaskByUidAndTaskId(String uid, String taskId) {
-        User user = userRepo.findByUid(uid);
-        checkUserInDB(user, uid);
-        return user.getTaskByTaskId(taskId);
-    }
-
-    private void checkUserInDB(User user, String identifier) {
-        if (user == null) {
-            log.error("User '{}' not found in database", identifier);
-            throw new EntityNotFoundException("User '" + identifier + "' not found in database");
-        } else {
-            log.info("User '{}' found in database", identifier);
+        if (!userRepo.existsById(uid)) {
+            throw new EntityNotFoundException("User not found");
         }
+        userRepo.deleteById(uid);
+        log.info("User '{}' deleted", uid);
+    }
+
+    /**
+     * Delete the user by email, throw exception if user not found
+     * @param email email
+     */
+    public void deleteByEmail(String email) {
+        if (!userRepo.existsByEmail(email)) {
+            throw new EntityNotFoundException("User not found");
+        }
+        userRepo.deleteByEmail(email);
+        log.info("User '{}' deleted", email);
     }
 }
